IncoQueryDjango/requirements.txt
django
mysql-connector-python
django-bower
django-nvd3
python-nvd3